# perf(dashboards): Increase dashboards list API page size